+++ generator/lib/src/generators/generator_registry.dart
@@ -1,9 +1,9 @@
-import 'package:forge/src/core/interfaces/component_generator.dart';
-import 'package:forge/src/generators/component/button_generator.dart';
-import 'package:forge/src/generators/component/input_generator.dart';
-import 'package:forge/src/generators/component/text_generator.dart';
-import 'package:forge/src/generators/component/generic_generator.dart';
-import 'package:forge/src/models/component_definition.dart';
+import 'package:syntax/src/core/interfaces/component_generator.dart';
+import 'package:syntax/src/generators/component/button_generator.dart';
+import 'package:syntax/src/generators/component/input_generator.dart';
+import 'package:syntax/src/generators/component/text_generator.dart';
+import 'package:syntax/src/generators/component/generic_generator.dart';
+import 'package:syntax/src/models/component_definition.dart';
 
 /// Registry of component generators.
 ///