+++ generator/lib/src/generators/registry/icon_registry_generator.dart
@@ -1,5 +1,5 @@
 import 'package:code_builder/code_builder.dart';
-import 'package:forge/src/parser/registry_parser.dart';
+import 'package:syntax/src/parser/registry_parser.dart';
 
 class IconRegistryGenerator {
   IconRegistryGenerator();