+++ generator/test/parser/enum_parsing_test.dart
@@ -2,7 +2,7 @@ import 'dart:io';
 
 import 'package:test/test.dart';
 import 'package:mason_logger/mason_logger.dart';
-import 'package:forge/src/parser/meta_parser.dart';
+import 'package:syntax/src/parser/meta_parser.dart';
 
 void main() {
   group('MetaParser Enum Parsing', () {
@@ -23,9 +23,9 @@ void main() {
     test('parses enums from meta file', () async {
       final file = File('${tempDir.path}/button.meta.dart');
       await file.writeAsString('''
-import 'package:forge/forge.dart';
+import 'package:syntax/syntax.dart';
 
-@ForgeComponent
+@SyntaxComponent
 class ButtonMeta {
   final String label;
   final ButtonVariant variant;