+++ generator/CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 0.0.1
 
 - Initial Alpha Release.
-- Added `Forge` CLI with `build` and `clean` commands.
+- Added `Syntax` CLI with `build` and `clean` commands.
 - Implemented `Button` component generator.
 - Added support for multiple themes (`Material`, `Cupertino`, `Neo`).
 - Implemented 5-layer SOLID architecture.